refactor(interpreter): extract realm state + statement handlers into partials (#1142)
Pure mechanical partial-class split of the 2,898-line Interpreter.cs, no
behaviour change:

- New Interpreter.Realm.cs holds the realm/global-state cluster: the process-
  wide globals table (CreateGlobalsLookup + the ordering-sensitive Promise/
  RegExp static trio, kept together and in textual order), the per-realm
  Symbol.for registry, Math, primitive prototypes, RegExp.prototype, and the
  globalThis routing helpers.
- The statement dispatch core and all Visit* handlers move next to their
  Execute* bodies in Interpreter.Statements.cs.

Interpreter.cs shrinks 2,898 -> 1,977 lines. dotnet test 15404/0.

Part of #1094.
