+++ Execution/Interpreter.cs
@@ -28,8 +28,9 @@ namespace SharpTS.Execution;
 ///
 /// This class is split across multiple partial class files:
 /// <list type="bullet">
-///   <item><description>Interpreter.cs - Core infrastructure and statement dispatch</description></item>
-///   <item><description>Interpreter.Statements.cs - Statement execution helpers (block, switch, try/catch, loops)</description></item>
+///   <item><description>Interpreter.cs - Core infrastructure, event loop, module execution</description></item>
+///   <item><description>Interpreter.Realm.cs - Realm/global state (globals table, symbol registry, primitive prototypes, globalThis routing)</description></item>
+///   <item><description>Interpreter.Statements.cs - Statement dispatch, Visit* handlers, and execution helpers (block, switch, try/catch, loops)</description></item>
 ///   <item><description>Interpreter.Expressions.cs - Expression dispatch and basic evaluators</description></item>
 ///   <item><description>Interpreter.Properties.cs - Property/member access (Get, Set, New, This)</description></item>
 ///   <item><description>Interpreter.Calls.cs - Function calls and binary/logical operators</description></item>
@@ -46,200 +47,6 @@ public partial class Interpreter : IDisposable
     /// </summary>
     private static readonly NodeRegistry<Interpreter, RuntimeValue, ExecutionResult> _registry =
         InterpreterRegistry.Create();
-
-    /// <summary>
-    /// The %Promise% constructor sentinel registered when no Promise singleton
-    /// claimed the global first. MUST be declared before
-    /// <see cref="_globalConstants"/> — static initializers run in textual
-    /// order and CreateGlobalsLookup reads this field.
-    /// </summary>
-    internal static readonly SharpTSBuiltInConstructor PromiseConstructorSentinel = new(
-        BuiltInNames.Promise,
-        _ => throw new Exception("Runtime Error: Use 'new Promise(executor)' syntax."));
-
-    /// <summary>
-    /// Frozen dictionary of global constants and built-in singletons for fast lookup.
-    /// Combines global constants (NaN, Infinity, undefined) with built-in namespaces
-    /// (Math, JSON, Object, etc.) into a single lookup to minimize dictionary operations.
-    /// </summary>
-    private static readonly FrozenDictionary<string, object> _globalConstants = CreateGlobalsLookup();
-
-    // The process-wide RegExp constructor singleton (a SharpTSBuiltInConstructor),
-    // resolved once from the static globals table. ECMA-262 §22.2.6.1 requires
-    // `RegExp.prototype.constructor === RegExp` and, by inheritance,
-    // `(/x/).constructor === RegExp` — both must reference this exact instance
-    // for strict-equality identity to hold. Cached so the regex property hot
-    // path returns it without a dictionary probe. Mirrors the compiled side,
-    // where the `$RegExp` Type token plays the same role.
-    internal static readonly object? RegExpConstructorObject =
-        _globalConstants.TryGetValue(BuiltInNames.RegExp, out var rxCtor) ? rxCtor : null;
-
-    private static FrozenDictionary<string, object> CreateGlobalsLookup()
-    {
-        var globals = new Dictionary<string, object>
-        {
-            [BuiltInNames.NaN] = double.NaN,
-            [BuiltInNames.Infinity] = double.PositiveInfinity,
-            [BuiltInNames.Undefined] = Runtime.Types.SharpTSUndefined.Instance,
-            [BuiltInNames.Fetch] = Runtime.Types.SharpTSFetchGlobal.Instance,
-
-            // SharedArrayBuffer constructor
-            [BuiltInNames.SharedArrayBuffer] = WorkerBuiltIns.SharedArrayBufferConstructor,
-
-            // ArrayBuffer constructor
-            [BuiltInNames.ArrayBuffer] = WorkerBuiltIns.ArrayBufferConstructor,
-
-            // DataView constructor
-            [BuiltInNames.DataView] = WorkerBuiltIns.DataViewConstructor,
-        };
-
-        // Add TypedArray constructors using centralized names
-        foreach (var typedArrayName in BuiltInNames.TypedArrayNames)
-        {
-            globals[typedArrayName] = WorkerBuiltIns.GetTypedArrayConstructor(typedArrayName);
-        }
-
-        // Add Error constructors as global class variables
-        // This enables typeof Error, class MyError extends Error, const E = Error, etc.
-        var errorClass = new Runtime.Types.SharpTSErrorClass("Error", null);
-        globals[BuiltInNames.Error] = errorClass;
-        foreach (var errorTypeName in BuiltInNames.ErrorTypeNames)
-        {
-            if (errorTypeName != "Error")
-                globals[errorTypeName] = new Runtime.Types.SharpTSErrorClass(errorTypeName, errorClass);
-        }
-
-        // Bare `Array` reference — needed for Array.prototype.X.apply() patterns
-        // that real-world CJS packages (yaml, lodash internals) rely on.
-        globals[BuiltInNames.Array] = Runtime.Types.SharpTSArrayGlobal.Instance;
-
-        // Bare `Function` reference — required for `Function.prototype.call.bind(...)`
-        // patterns used by test262 propertyHelper.js (and many libraries' native-
-        // detection paths). Without this, the harness fails at load before any
-        // test body runs.
-        globals[BuiltInNames.Function] = Runtime.Types.SharpTSFunctionGlobal.Instance;
-
-        // Node-style `global` alias for globalThis. CJS packages (lodash)
-        // detect the global object via `typeof global == 'object'` and alias
-        // its Array/Object/Date/etc. into a local scope.
-        var gtSingleton = BuiltInRegistry.Instance.GetSingleton(BuiltInNames.GlobalThis);
-        if (gtSingleton != null)
-        {
-            globals["global"] = gtSingleton;
-        }
-
-        // Add built-in singletons (Math, JSON, Object, etc.)
-        // These are namespaces that resolve to singleton instances when accessed as variables
-        string[] singletonNames =
-        [
-            BuiltInNames.Math, BuiltInNames.JSON, BuiltInNames.Object,
-            BuiltInNames.Number, BuiltInNames.String, BuiltInNames.Boolean, BuiltInNames.Symbol,
-            BuiltInNames.Console, BuiltInNames.Process, BuiltInNames.GlobalThis,
-            BuiltInNames.Reflect, BuiltInNames.Promise, BuiltInNames.Atomics,
-            "Buffer",
-            // WebCrypto global (#1063): bare `crypto` — an import binding shadows it.
-            "crypto",
-        ];
-        foreach (var name in singletonNames)
-        {
-            var singleton = BuiltInRegistry.Instance.GetSingleton(name);
-            if (singleton != null)
-            {
-                globals[name] = singleton;
-            }
-        }
-
-        // Add built-in constructors as global variables (Map, Set, Date, RegExp, etc.)
-        // Enables typeof Map, val instanceof Map, passing Map as value, Map.groupBy(), etc.
-        foreach (var (name, factory) in BuiltInConstructorFactory.GetConstructors())
-        {
-            if (!globals.ContainsKey(name))
-                globals[name] = new SharpTSBuiltInConstructor(name, factory);
-        }
-
-        // Expose global functions (parseFloat, parseInt, isNaN, isFinite,
-        // structuredClone, setTimeout/clearTimeout, etc.) as first-class
-        // callable values so they can be referenced by name — not just
-        // invoked directly. CommonJS packages (lodash) alias `var
-        // freeParseFloat = parseFloat`, and user code may do
-        // `typeof parseFloat === 'function'`.
-        string[] globalFunctionNames =
-        [
-            BuiltInNames.ParseInt, BuiltInNames.ParseFloat,
-            BuiltInNames.IsNaN, BuiltInNames.IsFinite,
-            BuiltInNames.StructuredClone,
-            BuiltInNames.EncodeURIComponent, BuiltInNames.DecodeURIComponent,
-            BuiltInNames.SetTimeout, BuiltInNames.ClearTimeout,
-            BuiltInNames.SetInterval, BuiltInNames.ClearInterval,
-            BuiltInNames.QueueMicrotask,
-        ];
-        foreach (var name in globalFunctionNames)
-        {
-            if (!globals.ContainsKey(name))
-                globals[name] = new SharpTSGlobalFunction(name);
-        }
-
-        // Bind value-position globals for built-ins that were previously only
-        // reachable through special-cased `new` expressions or member access
-        // (#208): bare `AbortSignal`/`Intl`/`ReadableStream`/... otherwise
-        // throw "Undefined variable".
-        //
-        // AbortSignal and Intl are namespace-style globals: member access on
-        // SharpTSBuiltInConstructor routes through the namespace registry
-        // (AbortSignal.abort/timeout/any, Intl.NumberFormat/...), while
-        // direct construction throws per spec (AbortSignal has no public
-        // constructor; Intl is not a constructor).
-        globals["AbortSignal"] = new SharpTSBuiltInConstructor("AbortSignal",
-            _ => throw new Exception("Runtime Error: TypeError: AbortSignal cannot be constructed directly. Use AbortSignal.abort(), AbortSignal.timeout(), or AbortController."));
-        globals["Intl"] = new SharpTSBuiltInConstructor("Intl",
-            _ => throw new Exception("Runtime Error: TypeError: Intl is not a constructor."));
-
-        // Web-streams constructors: the same singletons stream/web exports,
-        // so `new ReadableStream(...)`, `ReadableStream.from(...)`, and
-        // value-position references all share one identity.
-        globals[BuiltInNames.ReadableStream] = Runtime.Types.SharpTSReadableStreamConstructor.Instance;
-        globals[BuiltInNames.WritableStream] = Runtime.Types.SharpTSWritableStreamConstructor.Instance;
-        globals[BuiltInNames.TransformStream] = Runtime.Types.SharpTSTransformStreamConstructor.Instance;
-
-        // MessageChannel as a value (construction already worked by name).
-        globals[BuiltInNames.MessageChannel] = WorkerBuiltIns.MessageChannelConstructor;
-
-        // Symbol as a value-position global (#234): `typeof Symbol`,
-        // `const f = Symbol`, and `(Symbol as any).species` need a real
-        // binding. Its namespace is registered as non-singleton (member
-        // access routes through SymbolBuiltIns via GetMember), so the
-        // singleton loop above didn't bind it. The factory implements the
-        // call form Symbol(description); JS has no `new Symbol()`.
-        globals[BuiltInNames.Symbol] = new SharpTSBuiltInConstructor(
-            BuiltInNames.Symbol,
-            args => new SharpTSSymbol(args.Count > 0 && args[0] is not SharpTSUndefined
-                ? args[0]?.ToString()
-                : null));
-
-        // Promise needs a bare-reference global so `x instanceof Promise`,
-        // `typeof Promise === 'function'`, and stdlib modules that carry
-        // Promise as a value can type-check/run. Its namespace is registered
-        // as non-singleton (to preserve special `new Promise(executor)`
-        // handling), so it wasn't picked up by the loops above. Register a
-        // minimal constructor sentinel — `new Promise(executor)` has its
-        // own dedicated path and does not route through this factory.
-        if (!globals.ContainsKey(BuiltInNames.Promise))
-        {
-            globals[BuiltInNames.Promise] = PromiseConstructorSentinel;
-        }
-
-        return globals.ToFrozenDictionary();
-    }
-
-    /// <summary>
-    /// The value bare <c>Promise</c> resolves to — whatever the global table
-    /// actually holds (a registry singleton when one exists, otherwise
-    /// <see cref="PromiseConstructorSentinel"/>). Surfaced as
-    /// <c>promise.constructor</c> by PromiseBuiltIns.GetMember so the
-    /// ECMA-262 §27.2.5.1 identity holds:
-    /// <c>Promise.resolve(1).constructor === Promise</c> (#221).
-    /// </summary>
-    internal static object PromiseGlobalValue => _globalConstants[BuiltInNames.Promise];
 
     private RuntimeEnvironment _environment = new();
     // Keyed by AST-node identity, not structural value. The resolver stores and reads the
@@ -323,158 +130,6 @@ public partial class Interpreter : IDisposable
         _syncContext = new SyncEvaluationContext(this);
         _asyncContext = new AsyncEvaluationContext(this);
     }
-
-    // Per-realm RegExp.prototype. Held on the Interpreter (not on the
-    // process-wide SharpTSBuiltInConstructor singleton) so user mutations
-    // — `delete RegExp.prototype[Symbol.split]`, `Object.defineProperty`,
-    // etc. — stay scoped to this realm. Lazily populated on first read of
-    // `RegExp.prototype`.
-    private Runtime.Types.SharpTSObject? _regExpPrototype;
-    internal Runtime.Types.SharpTSObject GetRegExpPrototype()
-        => _regExpPrototype ??= Runtime.BuiltIns.RegExpBuiltIns.BuildPrototype();
-
-    // Per-realm Symbol.for registry. Held on the Interpreter (not as a
-    // process-wide static on SharpTSSymbol) so `Symbol.for(k)` returns a
-    // symbol unique to this realm and `Symbol.keyFor` cannot leak
-    // registrations across Interpreter instances. Each realm is its own agent
-    // per ECMA-262, so a separate registry is the correct semantics — and it
-    // removes a cross-thread data race: the old static was a plain Dictionary
-    // mutated by every realm in the process, including concurrent worker
-    // threads. Mirrors the per-realm RegExp.prototype (#101). Well-known
-    // symbols (Symbol.iterator, …) are NOT in this registry; they remain
-    // process-wide singletons. Lazily allocated; thread-confined to this
-    // realm's execution thread, so no lock is needed.
-    private Dictionary<string, Runtime.Types.SharpTSSymbol>? _symbolRegistry;
-    private Dictionary<Runtime.Types.SharpTSSymbol, string>? _symbolReverseRegistry;
-
-    /// <summary>
-    /// Returns this realm's registered symbol for <paramref name="key"/>,
-    /// creating and registering one on first use (ECMA-262 <c>Symbol.for</c>).
-    /// </summary>
-    internal Runtime.Types.SharpTSSymbol SymbolFor(string key)
-    {
-        _symbolRegistry ??= [];
-        if (_symbolRegistry.TryGetValue(key, out var existing))
-            return existing;
-
-        var symbol = new Runtime.Types.SharpTSSymbol(key);
-        _symbolRegistry[key] = symbol;
-        (_symbolReverseRegistry ??= [])[symbol] = key;
-        return symbol;
-    }
-
-    /// <summary>
-    /// Returns the registry key for <paramref name="symbol"/> in this realm, or
-    /// <c>null</c> if it was not produced by this realm's <c>Symbol.for</c>
-    /// (ECMA-262 <c>Symbol.keyFor</c>).
-    /// </summary>
-    internal string? SymbolKeyFor(Runtime.Types.SharpTSSymbol symbol)
-        => _symbolReverseRegistry is not null
-            && _symbolReverseRegistry.TryGetValue(symbol, out var key)
-            ? key
-            : null;
-
-    // Per-realm Math. Math is an extensible ECMA-262 object: guest code may add
-    // properties (`Math.x = 1`), which must not leak across realms or race
-    // across worker threads. Held per-Interpreter, mirroring RegExp.prototype
-    // (#101). The base members (PI, sqrt, …) are stateless and resolved the
-    // same way for every instance; only the per-instance `_extras` overlay
-    // differs. Within a realm both the bare `Math` global and `globalThis.Math`
-    // resolve to this one instance, so `Math === globalThis.Math` holds.
-    private Runtime.Types.SharpTSMath? _math;
-    internal Runtime.Types.SharpTSMath GetMath() => _math ??= new Runtime.Types.SharpTSMath();
-
-    /// <summary>
-    /// Resolves a per-realm mutable built-in intrinsic by its global name
-    /// (currently <c>Math</c>). These are the built-ins moved off process-global
-    /// singletons so a realm's guest mutations stay realm-local. Returns
-    /// <c>false</c> for every other name, leaving normal global resolution
-    /// unchanged.
-    /// </summary>
-    internal bool TryGetRealmIntrinsic(string name, out object? value)
-    {
-        if (IsRealmIntrinsicName(name))
-        {
-            value = GetMath();
-            return true;
-        }
-        value = null;
-        return false;
-    }
-
-    /// <summary>
-    /// Names of the per-realm mutable built-ins resolved off the Interpreter
-    /// rather than the shared global-constants table or the namespace
-    /// fast-path. Used to keep all resolution routes (bare global, namespace
-    /// member access, <c>globalThis</c>) pointing at the one realm instance so
-    /// method identity holds (<c>Math.max === Math.max</c>).
-    /// </summary>
-    internal static bool IsRealmIntrinsicName(string name) => name == "Math";
-
-    // Per-realm String/Number/Boolean.prototype. Each is an extensible ECMA-262
-    // object carrying a guest-writable _extras bag, so — like Math and
-    // RegExp.prototype (#101) — it is held per-Interpreter: guest writes
-    // (`String.prototype.x = …`, indexed/`length` assignments Test262 makes
-    // before calling Array.prototype.* on a primitive) stay realm-local and
-    // don't race across worker threads. The namespace objects
-    // (String/Number/Boolean themselves) are immutable and stay shared
-    // singletons; only the mutable prototypes are per-realm.
-    private Runtime.Types.SharpTSStringPrototype? _stringPrototype;
-    private Runtime.Types.SharpTSNumberPrototype? _numberPrototype;
-    private Runtime.Types.SharpTSBooleanPrototype? _booleanPrototype;
-    internal Runtime.Types.SharpTSStringPrototype GetStringPrototype() => _stringPrototype ??= new();
-    internal Runtime.Types.SharpTSNumberPrototype GetNumberPrototype() => _numberPrototype ??= new();
-    internal Runtime.Types.SharpTSBooleanPrototype GetBooleanPrototype() => _booleanPrototype ??= new();
-
-    // Per-realm globalThis. The global object holds guest-assigned properties
-    // (`globalThis.x = …`), which must stay realm-local and not race across
-    // worker threads, so each Interpreter owns its own — like RegExp.prototype
-    // (#101), Math, and the primitive prototypes. Built-in namespaces (Math,
-    // JSON, …) are still resolved live through the shared BuiltInRegistry, so
-    // `globalThis.JSON === JSON` and `globalThis.Math === Math` still hold; only
-    // the user-property bag is per-realm. Bare `globalThis` and the Node
-    // `global` alias resolve here (see LookupVariableRV), and sloppy-mode `this`
-    // binds to it.
-    private Runtime.Types.SharpTSGlobalThis? _globalThis;
-    internal Runtime.Types.SharpTSGlobalThis GlobalThis => _globalThis ??= new Runtime.Types.SharpTSGlobalThis();
-
-    /// <summary>
-    /// Resolves <c>String</c>/<c>Number</c>/<c>Boolean</c><c>.prototype</c> to
-    /// this realm's prototype instance when <paramref name="obj"/> is the
-    /// corresponding built-in namespace, so the read is realm-local rather than
-    /// the shared singleton. Returns <c>false</c> for any other receiver,
-    /// leaving normal member resolution unchanged.
-    /// </summary>
-    private bool TryGetRealmPrototypeForNamespace(object? obj, out object? prototype)
-    {
-        switch (obj)
-        {
-            case Runtime.Types.SharpTSStringNamespace:
-                prototype = GetStringPrototype();
-                return true;
-            case Runtime.Types.SharpTSNumberNamespace:
-                prototype = GetNumberPrototype();
-                return true;
-            case Runtime.Types.SharpTSBooleanNamespace:
-                prototype = GetBooleanPrototype();
-                return true;
-            default:
-                prototype = null;
-                return false;
-        }
-    }
-
-    /// <summary>
-    /// Reads a property off <c>globalThis</c> honoring per-realm intrinsics: a
-    /// guest own-assignment (<c>globalThis.Math = x</c>) wins, then the realm
-    /// intrinsic (so <c>globalThis.Math === Math</c> within a realm), then the
-    /// normal built-in/global resolution. Behaviour is identical to
-    /// <c>globalThis.GetProperty</c> for every non-intrinsic name.
-    /// </summary>
-    private object? ResolveGlobalThisRead(Runtime.Types.SharpTSGlobalThis globalThis, string key)
-        => !globalThis.HasUserProperty(key) && TryGetRealmIntrinsic(key, out var intrinsic)
-            ? intrinsic
-            : globalThis.GetProperty(key);
 
     // Module support
     private readonly Dictionary<string, ModuleInstance> _loadedModules = [];
@@ -2319,747 +1974,5 @@ public partial class Interpreter : IDisposable
         };
         return _environment.Get(token).ToObject();
     }
-
-    /// <summary>
-    /// Internal wrapper for Execute that allows evaluation contexts to dispatch statements.
-    /// </summary>
-    /// <param name="stmt">The statement to execute.</param>
-    /// <returns>The execution result.</returns>
-    internal ExecutionResult ExecuteStatement(Stmt stmt) => Execute(stmt);
-
-    /// <summary>
-    /// Internal async wrapper for statement execution using registry-based dispatch.
-    /// Uses DispatchStmtAsync which falls back to sync handlers when no async handler exists.
-    /// </summary>
-    /// <param name="stmt">The statement to execute.</param>
-    /// <returns>A task containing the execution result.</returns>
-    internal async Task<ExecutionResult> ExecuteStatementAsync(Stmt stmt)
-    {
-        return await _registry.DispatchStmtAsync(stmt, this);
-    }
-
-    /// <summary>
-    /// Dispatches a statement to the appropriate execution handler using the registry.
-    /// </summary>
-    /// <param name="stmt">The statement AST node to execute.</param>
-    /// <remarks>
-    /// Handles all statement types including control flow (if, while, for, switch),
-    /// declarations (var, function, class, enum), and control transfer (return, break, continue, throw).
-    /// Control flow uses <see cref="ExecutionResult"/> for non-local jumps.
-    /// </remarks>
-    private ExecutionResult Execute(Stmt stmt)
-    {
-        return _registry.DispatchStmt(stmt, this);
-    }
-
-    // Statement handlers - called by the registry
-
-    internal ExecutionResult VisitBlock(Stmt.Block block) =>
-        ExecuteBlock(block.Statements, new RuntimeEnvironment(_environment));
-
-    internal ExecutionResult VisitLabeledStatement(Stmt.LabeledStatement labeledStmt) =>
-        ExecuteLabeledStatement(labeledStmt);
-
-    internal ExecutionResult VisitSequence(Stmt.Sequence seq)
-    {
-        // Execute in current scope (no new environment)
-        foreach (var s in seq.Statements)
-        {
-            var result = Execute(s);
-            if (result.IsAbrupt) return result;
-        }
-        return ExecutionResult.Success();
-    }
-
-    internal ExecutionResult VisitExpression(Stmt.Expression exprStmt)
-    {
-        Evaluate(exprStmt.Expr);
-        return ExecutionResult.Success();
-    }
-
-    internal ExecutionResult VisitIf(Stmt.If ifStmt)
-    {
-        if (IsTruthy(Evaluate(ifStmt.Condition)))
-        {
-            return Execute(ifStmt.ThenBranch);
-        }
-        else if (ifStmt.ElseBranch != null)
-        {
-            return Execute(ifStmt.ElseBranch);
-        }
-        return ExecutionResult.Success();
-    }
-
-    internal ExecutionResult VisitWhile(Stmt.While whileStmt)
-    {
-        var labels = TakePendingLoopLabels();
-        return ExecuteWhileCore(_syncContext, whileStmt, labels).GetAwaiter().GetResult();
-    }
-
-    internal ExecutionResult VisitDoWhile(Stmt.DoWhile doWhileStmt)
-    {
-        var labels = TakePendingLoopLabels();
-        return ExecuteDoWhileCore(_syncContext, doWhileStmt, labels).GetAwaiter().GetResult();
-    }
-
-    internal ExecutionResult VisitFor(Stmt.For forStmt)
-    {
-        // Drain labels parked by an enclosing labeled statement before running the initializer,
-        // so `continue <label>`/`break <label>` resolve to this loop (#558).
-        var labels = TakePendingLoopLabels();
-        return ExecuteForCore(_syncContext, forStmt, labels).GetAwaiter().GetResult();
-    }
-
-    /// <summary>
-    /// Core C-style for-loop execution logic shared by the sync and async evaluators. The evaluation
-    /// context supplies the per-clause evaluation strategy and the between-iteration scheduler yield
-    /// (<see cref="IEvaluationContext.YieldToSchedulerAsync"/>), so a single body serves both paths.
-    /// </summary>
-    private async ValueTask<ExecutionResult> ExecuteForCore(IEvaluationContext ctx, Stmt.For forStmt, IReadOnlyList<string>? labels)
-    {
-        // Create scope for loop variables (ES6 let/const block scoping)
-        // Variables declared with let/const in the initializer are scoped to the loop
-        RuntimeEnvironment loopEnv = new(_environment);
-        using (PushScope(loopEnv))
-        {
-            // Execute initializer once (defines loop variable in loopEnv)
-            if (forStmt.Initializer != null)
-                await ctx.ExecuteStmtAsync(forStmt.Initializer);
-            // ECMA-262 13.7.4: a `for (let/const …)` loop gives each iteration its
-            // own bindings for the loop variables, so closures created in different
-            // iterations capture distinct values (#633). `var`/expression
-            // initializers share a single binding and keep the no-copy fast path.
-            var perIterationNames = CollectPerIterationBindings(forStmt.Initializer);
-            if (perIterationNames != null)
-                CreatePerIterationEnvironment(loopEnv, perIterationNames);
-            // Loop with proper continue handling - increment always runs
-            while (forStmt.Condition == null || (await ctx.EvaluateExprAsync(forStmt.Condition)).IsTruthy())
-            {
-                var result = await ctx.ExecuteStmtAsync(forStmt.Body);
-                var (shouldBreak, shouldContinue, abruptResult) = HandleLoopResult(result, labels);
-                if (shouldBreak) break;
-                // On continue (unlabeled or to this loop), execute increment then re-test
-                if (shouldContinue)
-                {
-                    if (perIterationNames != null)
-                        CreatePerIterationEnvironment(loopEnv, perIterationNames);
-                    if (forStmt.Increment != null)
-                        await ctx.EvaluateExprAsync(forStmt.Increment);
-                    // Yield to allow timer callbacks and other threads to execute
-                    await ctx.YieldToSchedulerAsync();
-                    continue;
-                }
-                if (abruptResult.HasValue) return abruptResult.Value;
-                // Normal completion: fresh per-iteration binding, then increment
-                if (perIterationNames != null)
-                    CreatePerIterationEnvironment(loopEnv, perIterationNames);
-                if (forStmt.Increment != null)
-                    await ctx.EvaluateExprAsync(forStmt.Increment);
-                // Process any pending timer callbacks
-                ProcessPendingCallbacks();
-            }
-            return ExecutionResult.Success();
-        }
-    }
-
-    /// <summary>
-    /// Returns the variable names a <c>for</c> initializer binds that require a
-    /// fresh binding per iteration (ECMA-262 13.7.4): <c>let</c>/<c>const</c>
-    /// declarations. Returns <c>null</c> for <c>var</c> or expression
-    /// initializers, which share a single binding across the whole loop.
-    /// </summary>
-    private static List<string>? CollectPerIterationBindings(Stmt? initializer)
-    {
-        switch (initializer)
-        {
-            // `let`/`const` in a for-initializer parse to Stmt.Var with IsVar=false.
-            case Stmt.Var v when !v.IsVar:
-                return [v.Name.Lexeme];
-            case Stmt.Const c:
-                return [c.Name.Lexeme];
-            // Multi-declarator initializers (`for (let i = 0, j = 10; …)`).
-            case Stmt.Sequence seq:
-                List<string>? names = null;
-                foreach (var s in seq.Statements)
-                {
-                    var sub = CollectPerIterationBindings(s);
-                    if (sub != null) (names ??= []).AddRange(sub);
-                }
-                return names;
-            default:
-                return null;
-        }
-    }
-
-    /// <summary>
-    /// ECMA-262 13.7.4.8 CreatePerIterationEnvironment: copies the current value
-    /// of each loop variable into a fresh environment that is a sibling of the
-    /// loop environment (same enclosing scope, so the resolver's static scope
-    /// distances stay valid) and makes it the active scope. Closures created in
-    /// the next iteration capture this distinct binding rather than a shared slot.
-    /// </summary>
-    private void CreatePerIterationEnvironment(RuntimeEnvironment loopEnv, List<string> names)
-    {
-        var iterationEnv = new RuntimeEnvironment(loopEnv.Enclosing);
-        foreach (var name in names)
-            iterationEnv.Define(name, _environment.GetAt(0, name));
-        _environment = iterationEnv;
-    }
-
-    internal ExecutionResult VisitForOf(Stmt.ForOf forOf) => ExecuteForOf(forOf);
-
-    internal ExecutionResult VisitForIn(Stmt.ForIn forIn) => ExecuteForIn(forIn);
-
-    internal ExecutionResult VisitBreak(Stmt.Break breakStmt) =>
-        ExecutionResult.Break(breakStmt.Label?.Lexeme);
-
-    internal ExecutionResult VisitContinue(Stmt.Continue continueStmt) =>
-        ExecutionResult.Continue(continueStmt.Label?.Lexeme);
-
-    internal ExecutionResult VisitSwitch(Stmt.Switch switchStmt) => ExecuteSwitch(switchStmt);
-
-    internal ExecutionResult VisitTryCatch(Stmt.TryCatch tryCatch) => ExecuteTryCatch(tryCatch);
-
-    internal ExecutionResult VisitThrow(Stmt.Throw throwStmt) =>
-        ExecutionResult.Throw(Evaluate(throwStmt.Value));
-
-    internal ExecutionResult VisitVar(Stmt.Var varStmt)
-    {
-        object? value = SharpTSUndefined.Instance;
-        if (varStmt.Initializer != null)
-        {
-            value = Evaluate(varStmt.Initializer);
-        }
-        _environment.Define(varStmt.Name.Lexeme, value);
-        return ExecutionResult.Success();
-    }
-
-    internal ExecutionResult VisitConst(Stmt.Const constStmt)
-    {
-        // Const declarations always have an initializer (enforced by parser)
-        object? constValue = Evaluate(constStmt.Initializer);
-        _environment.Define(constStmt.Name.Lexeme, constValue);
-        return ExecutionResult.Success();
-    }
-
-    internal ExecutionResult VisitFunction(Stmt.Function functionStmt)
-    {
-        // Skip overload signatures (no body) - they're type-checking only
-        if (functionStmt.Body == null) return ExecutionResult.Success();
-        // Skip if already hoisted
-        if (_environment.IsDefinedLocally(functionStmt.Name.Lexeme)) return ExecutionResult.Success();
-        if (functionStmt.IsGenerator && functionStmt.IsAsync)
-        {
-            // Async generator: async function* foo() { yield await ... }
-            SharpTSAsyncGeneratorFunction asyncGenFunction = new(functionStmt, _environment);
-            _environment.Define(functionStmt.Name.Lexeme, asyncGenFunction);
-        }
-        else if (functionStmt.IsGenerator)
-        {
-            SharpTSGeneratorFunction generatorFunction = new(functionStmt, _environment);
-            _environment.Define(functionStmt.Name.Lexeme, generatorFunction);
-        }
-        else if (functionStmt.IsAsync)
-        {
-            SharpTSAsyncFunction asyncFunction = new(functionStmt, _environment);
-            _environment.Define(functionStmt.Name.Lexeme, asyncFunction);
-        }
-        else
-        {
-            SharpTSFunction function = new(functionStmt, _environment);
-            _environment.Define(functionStmt.Name.Lexeme, function);
-        }
-        return ExecutionResult.Success();
-    }
-
-    internal ExecutionResult VisitClass(Stmt.Class classStmt)
-    {
-        // @DotNetType declare class: bind a DotNet wrapper into the environment
-        // instead of creating an empty SharpTSClass. Non-DotNet declare classes still
-        // fall through and produce an empty SharpTSClass for type-only compatibility.
-        if (classStmt.IsDeclare)
-        {
-            if (TryRegisterDotNetType(classStmt)) return ExecutionResult.Success();
-        }
-
-        object? superclass = null;
-        if (classStmt.SuperclassExpr != null)
-        {
-            superclass = Evaluate(classStmt.SuperclassExpr);
-
-            // `extends Array` (#233): the Array global is a constructor
-            // singleton, not a SharpTSClass — substitute the SharpTSArrayClass
-            // bridge so the class machinery (super(), method lookup,
-            // instanceof) sees a real superclass.
-            if (superclass is SharpTSArrayGlobal)
-            {
-                superclass = SharpTSArrayClass.ArrayBase;
-            }
-
-            // `extends Promise` (#242): same substitution for the Promise
-            // constructor sentinel.
-            if (superclass is SharpTSBuiltInConstructor { Name: BuiltInNames.Promise })
-            {
-                superclass = SharpTSPromiseClass.PromiseBase;
-            }
-
-            if (superclass is not SharpTSClass)
-            {
-                // Built-in constructors that don't have a class bridge yet
-                // get a precise error instead of the generic
-                // "Superclass must be a class".
-                if (superclass is SharpTSBuiltInConstructor builtInCtor)
-                {
-                    throw new InterpreterException(
-                        $"Class '{classStmt.Name.Lexeme}' cannot extend built-in '{builtInCtor.Name}': subclassing this built-in is not supported yet.");
-                }
-                throw new InterpreterException("Superclass must be a class.");
-            }
-        }
-
-        _environment.Define(classStmt.Name.Lexeme, null);
-
-        if (classStmt.SuperclassExpr != null)
-        {
-            _environment = new RuntimeEnvironment(_environment);
-            _environment.Define("super", superclass);
-        }
-
-        Dictionary<string, ISharpTSCallable> methods = [];
-        Dictionary<string, ISharpTSCallable> staticMethods = [];
-        Dictionary<string, object?> staticProperties = [];
-        List<Stmt.Field> instanceFields = [];
-        // ES2022 private class elements
-        List<Stmt.Field> instancePrivateFields = [];
-        Dictionary<string, ISharpTSCallable> privateMethods = [];
-        Dictionary<string, object?> staticPrivateFields = [];
-        Dictionary<string, ISharpTSCallable> staticPrivateMethods = [];
-
-        // Process fields: collect instance fields, defer static field initialization if using StaticInitializers
-        // Note: Declare fields are processed normally - they can't have initializers (enforced by parser),
-        // so they'll be added with null/undefined values and can be set externally later.
-        bool hasStaticInitializers = classStmt.StaticInitializers != null && classStmt.StaticInitializers.Count > 0;
-
-        foreach (Stmt.Field field in classStmt.Fields)
-        {
-            if (field.IsPrivate)
-            {
-                // ES2022 private fields
-                if (field.IsStatic)
-                {
-                    if (!hasStaticInitializers)
-                    {
-                        // Old behavior: evaluate immediately
-                        object? fieldValue = field.Initializer != null
-                            ? Evaluate(field.Initializer)
-                            : null;
-                        staticPrivateFields[field.Name.Lexeme] = fieldValue;
-                    }
-                    // else: will be evaluated via StaticInitializers with proper 'this' binding
-                }
-                else
-                {
-                    // Collect instance private fields - they'll be initialized when instances are created
-                    instancePrivateFields.Add(field);
-                }
-            }
-            else if (field.IsStatic)
-            {
-                if (!hasStaticInitializers)
-                {
-                    // Old behavior: evaluate immediately
-                    object? fieldValue = field.Initializer != null
-                        ? Evaluate(field.Initializer)
-                        : null;
-                    staticProperties[field.Name.Lexeme] = fieldValue;
-                }
-                // else: will be evaluated via StaticInitializers with proper 'this' binding
-            }
-            else
-            {
-                // Collect instance fields - they'll be initialized when instances are created
-                instanceFields.Add(field);
-            }
-        }
-
-        // Symbol-keyed computed methods (`[Symbol.iterator]() {...}`) can't go into the string
-        // dictionaries; collected here and attached to the class after construction.
-        List<(SharpTSSymbol Symbol, ISharpTSCallable Func, bool IsStatic)>? symbolMethods = null;
-
-        // Separate static and instance methods (skip overload signatures with no body)
-        foreach (Stmt.Function method in classStmt.Methods.Where(m => m.Body != null))
-        {
-            // Create the appropriate function type based on async/generator flags
-            ISharpTSCallable func;
-            if (method.IsGenerator && method.IsAsync)
-                func = new SharpTSAsyncGeneratorFunction(method, _environment);
-            else if (method.IsAsync)
-                func = new SharpTSAsyncFunction(method, _environment);
-            else if (method.IsGenerator)
-                func = new SharpTSGeneratorFunction(method, _environment);
-            else
-                func = new SharpTSFunction(method, _environment);
-
-            // Computed method keys (`[Symbol.iterator]()`, `[expr]()`) are evaluated at
-            // class-definition time, like computed field keys and accessors. Symbol keys land
-            // in the symbol-method table; other keys fold to a string-named method.
-            if (method.ComputedKey != null)
-            {
-                object? key = Evaluate(method.ComputedKey);
-                if (key is SharpTSSymbol symbolKey)
-                {
-                    (symbolMethods ??= []).Add((symbolKey, func, method.IsStatic));
-                    continue;
-                }
-                string keyStr = PropertyKeyConverter.ToPropertyKeyString(key);
-                (method.IsStatic ? staticMethods : methods)[keyStr] = func;
-                continue;
-            }
-
-            if (method.IsPrivate)
-            {
-                // ES2022 private methods
-                if (method.IsStatic)
-                {
-                    staticPrivateMethods[method.Name.Lexeme] = func;
-                }
-                else
-                {
-                    privateMethods[method.Name.Lexeme] = func;
-                }
-            }
-            else if (method.IsStatic)
-            {
-                staticMethods[method.Name.Lexeme] = func;
-            }
-            else
-            {
-                methods[method.Name.Lexeme] = func;
-            }
-        }
-
-        // Create accessor functions
-        Dictionary<string, SharpTSFunction> getters = [];
-        Dictionary<string, SharpTSFunction> setters = [];
-        Dictionary<string, SharpTSFunction> staticGetters = [];
-        Dictionary<string, SharpTSFunction> staticSetters = [];
-
-        // Symbol-keyed accessors can't go into the string dictionaries; collected
-        // here and attached to the class after construction.
-        List<(SharpTSSymbol Symbol, SharpTSFunction Func, bool IsStatic, bool IsGetter)>? symbolAccessors = null;
-
-        if (classStmt.Accessors != null)
-        {
-            foreach (var accessor in classStmt.Accessors)
-            {
-                // Create a synthetic function for the accessor
-                var funcStmt = new Stmt.Function(
-                    accessor.Name,
-                    null,  // No type parameters for accessor
-                    null,  // No this type annotation
-                    accessor.SetterParam != null ? [accessor.SetterParam] : [],
-                    accessor.Body,
-                    accessor.ReturnType);
-
-                SharpTSFunction func = new(funcStmt, _environment);
-                bool isGetter = accessor.Kind.Type == TokenType.GET;
-
-                // Computed accessor names (`get [Symbol.toStringTag]()`,
-                // `static get [Symbol.species]()`) are evaluated at
-                // class-definition time, like computed field keys.
-                if (accessor.ComputedKey != null)
-                {
-                    object? key = Evaluate(accessor.ComputedKey);
-                    if (key is SharpTSSymbol symbolKey)
-                    {
-                        (symbolAccessors ??= []).Add((symbolKey, func, accessor.IsStatic, isGetter));
-                        continue;
-                    }
-                    string keyStr = PropertyKeyConverter.ToPropertyKeyString(key);
-                    if (isGetter) (accessor.IsStatic ? staticGetters : getters)[keyStr] = func;
-                    else (accessor.IsStatic ? staticSetters : setters)[keyStr] = func;
-                    continue;
-                }
-
-                var targetGet = accessor.IsStatic ? staticGetters : getters;
-                var targetSet = accessor.IsStatic ? staticSetters : setters;
-
-                if (isGetter)
-                {
-                    targetGet[accessor.Name.Lexeme] = func;
-                }
-                else
-                {
-                    targetSet[accessor.Name.Lexeme] = func;
-                }
-            }
-        }
-
-        // Process auto-accessors (TypeScript 4.9+)
-        List<Stmt.AutoAccessor> instanceAutoAccessors = [];
-        Dictionary<string, object?> staticAutoAccessors = [];
-
-        if (classStmt.AutoAccessors != null)
-        {
-            foreach (var autoAccessor in classStmt.AutoAccessors)
-            {
-                if (autoAccessor.IsStatic)
-                {
-                    // Evaluate static auto-accessor initializer now
-                    object? initValue = autoAccessor.Initializer != null
-                        ? Evaluate(autoAccessor.Initializer)
-                        : null;
-                    staticAutoAccessors[autoAccessor.Name.Lexeme] = initValue;
-                }
-                else
-                {
-                    // Collect instance auto-accessors for later initialization
-                    instanceAutoAccessors.Add(autoAccessor);
-                }
-            }
-        }
-
-        // If the superclass is an Error type, create a SharpTSErrorClass so that
-        // instances carry error fields (name, message, stack) and instanceof works.
-        // Likewise an Array superclass produces a SharpTSArrayClass whose
-        // instances are real arrays (#233), and a Promise superclass produces a
-        // SharpTSPromiseClass whose instances are real promises (#242).
-        SharpTSClass klass = superclass is SharpTSErrorClass errorSuper
-            ? new SharpTSErrorClass(
-                classStmt.Name.Lexeme,
-                errorSuper,
-                methods,
-                staticMethods,
-                staticProperties,
-                getters,
-                setters,
-                classStmt.IsAbstract,
-                instanceFields,
-                instancePrivateFields,
-                privateMethods,
-                staticPrivateFields,
-                staticPrivateMethods,
-                instanceAutoAccessors.Count > 0 ? instanceAutoAccessors : null,
-                staticAutoAccessors.Count > 0 ? staticAutoAccessors : null,
-                staticGetters.Count > 0 ? staticGetters : null,
-                staticSetters.Count > 0 ? staticSetters : null)
-            : superclass is SharpTSArrayClass arraySuper
-            ? new SharpTSArrayClass(
-                classStmt.Name.Lexeme,
-                arraySuper,
-                methods,
-                staticMethods,
-                staticProperties,
-                getters,
-                setters,
-                classStmt.IsAbstract,
-                instanceFields,
-                instancePrivateFields,
-                privateMethods,
-                staticPrivateFields,
-                staticPrivateMethods,
-                instanceAutoAccessors.Count > 0 ? instanceAutoAccessors : null,
-                staticAutoAccessors.Count > 0 ? staticAutoAccessors : null,
-                staticGetters.Count > 0 ? staticGetters : null,
-                staticSetters.Count > 0 ? staticSetters : null)
-            : superclass is SharpTSPromiseClass promiseSuper
-            ? new SharpTSPromiseClass(
-                classStmt.Name.Lexeme,
-                promiseSuper,
-                methods,
-                staticMethods,
-                staticProperties,
-                getters,
-                setters,
-                classStmt.IsAbstract,
-                instanceFields,
-                instancePrivateFields,
-                privateMethods,
-                staticPrivateFields,
-                staticPrivateMethods,
-                instanceAutoAccessors.Count > 0 ? instanceAutoAccessors : null,
-                staticAutoAccessors.Count > 0 ? staticAutoAccessors : null,
-                staticGetters.Count > 0 ? staticGetters : null,
-                staticSetters.Count > 0 ? staticSetters : null)
-            : new SharpTSClass(
-                classStmt.Name.Lexeme,
-                (SharpTSClass?)superclass,
-                methods,
-                staticMethods,
-                staticProperties,
-                getters,
-                setters,
-                classStmt.IsAbstract,
-                instanceFields,
-                instancePrivateFields,
-                privateMethods,
-                staticPrivateFields,
-                staticPrivateMethods,
-                instanceAutoAccessors.Count > 0 ? instanceAutoAccessors : null,
-                staticAutoAccessors.Count > 0 ? staticAutoAccessors : null,
-                staticGetters.Count > 0 ? staticGetters : null,
-                staticSetters.Count > 0 ? staticSetters : null);
-
-        if (symbolAccessors != null)
-        {
-            foreach (var (symbol, func, isStatic, isGetter) in symbolAccessors)
-            {
-                klass.AddSymbolAccessor(symbol, func, isStatic, isGetter);
-            }
-        }
-
-        if (symbolMethods != null)
-        {
-            foreach (var (symbol, func, isStatic) in symbolMethods)
-            {
-                klass.AddSymbolMethod(symbol, func, isStatic);
-            }
-        }
-
-        // Execute static initializers in declaration order (if present)
-        if (hasStaticInitializers)
-        {
-            // Create temporary environment with 'this' bound to the class
-            // Also make the class name available so code like Foo.x works
-            var staticEnv = new RuntimeEnvironment(_environment);
-            staticEnv.Define("this", klass);
-            staticEnv.Define(classStmt.Name.Lexeme, klass);
-
-            var prevEnv = _environment;
-            _environment = staticEnv;
-
-            try
-            {
-                foreach (var initializer in classStmt.StaticInitializers!)
-                {
-                    switch (initializer)
-                    {
-                        case Stmt.Field field when field.IsStatic:
-                            object? fieldValue = field.Initializer != null
-                                ? Evaluate(field.Initializer)
-                                : null;
-                            if (field.IsPrivate)
-                                klass.SetStaticPrivateField(field.Name.Lexeme, fieldValue);
-                            else
-                                klass.SetStaticProperty(field.Name.Lexeme, fieldValue);
-                            break;
-
-                        case Stmt.StaticBlock block:
-                            foreach (var blockStmt in block.Body)
-                            {
-                                var result = Execute(blockStmt);
-                                if (result.IsAbrupt)
-                                {
-                                    // Handle throw from static block
-                                    if (result.Type == ExecutionResult.ResultType.Throw)
-                                    {
-                                        throw new InterpreterException($"Error in static block: {Stringify(result.Value.ToObject())}");
-                                    }
-                                    // Return, break, continue are not allowed (validated by type checker)
-                                }
-                            }
-                            break;
-                    }
-                }
-            }
-            finally
-            {
-                _environment = prevEnv;
-            }
-        }
-
-        // Apply decorators in the correct order
-        klass = ApplyAllDecorators(classStmt, klass, methods, staticMethods, getters, setters);
-
-        if (classStmt.SuperclassExpr != null)
-        {
-            _environment = _environment.Enclosing!;
-        }
-
-        _environment.Assign(classStmt.Name, klass);
-        return ExecutionResult.Success();
-    }
-
-    internal ExecutionResult VisitTypeAlias(Stmt.TypeAlias typeAlias) =>
-        // Type-only declarations - compile-time only, no runtime effect
-        ExecutionResult.Success();
-
-    internal ExecutionResult VisitInterface(Stmt.Interface iface) =>
-        // Type-only declarations - compile-time only, no runtime effect
-        ExecutionResult.Success();
-
-    internal ExecutionResult VisitFileDirective(Stmt.FileDirective fileDirective) =>
-        // Type-only declarations - compile-time only, no runtime effect
-        ExecutionResult.Success();
-
-    internal ExecutionResult VisitField(Stmt.Field field) =>
-        // Class member declarations - handled within class processing, not executed directly
-        ExecutionResult.Success();
-
-    internal ExecutionResult VisitAccessor(Stmt.Accessor accessor) =>
-        // Class member declarations - handled within class processing, not executed directly
-        ExecutionResult.Success();
-
-    internal ExecutionResult VisitAutoAccessor(Stmt.AutoAccessor autoAccessor) =>
-        // Class member declarations - handled within class processing, not executed directly
-        ExecutionResult.Success();
-
-    internal ExecutionResult VisitStaticBlock(Stmt.StaticBlock staticBlock) =>
-        // Class member declarations - handled within class processing, not executed directly
-        ExecutionResult.Success();
-
-    internal ExecutionResult VisitEnum(Stmt.Enum enumStmt)
-    {
-        ExecuteEnumDeclaration(enumStmt);
-        return ExecutionResult.Success();
-    }
-
-    internal ExecutionResult VisitNamespace(Stmt.Namespace ns) => ExecuteNamespace(ns);
-
-    internal ExecutionResult VisitImportAlias(Stmt.ImportAlias importAlias) => ExecuteImportAlias(importAlias);
-
-    internal ExecutionResult VisitReturn(Stmt.Return returnStmt)
-    {
-        // A bare `return;` completes with `undefined` — distinct from `return null;`. Emitting the
-        // undefined sentinel here (rather than C# null, which represents JS null) is what makes a
-        // generator's completion value and a plain function's return value `undefined` instead of
-        // conflating them with null (#480). `return <expr>` preserves whatever the expression
-        // evaluates to, so an explicit `return null;` still yields null.
-        if (returnStmt.Value == null)
-            return ExecutionResult.Return(RuntimeValue.Undefined);
-        return ExecutionResult.Return(Evaluate(returnStmt.Value));
-    }
-
-    internal ExecutionResult VisitPrint(Stmt.Print printStmt)
-    {
-        Out.WriteLine(Stringify(Evaluate(printStmt.Expr)));
-        return ExecutionResult.Success();
-    }
-
-    internal ExecutionResult VisitImport(Stmt.Import import) =>
-        // Imports are handled in BindModuleImports before execution
-        // In single-file mode, imports are a no-op (type checker would have errored)
-        ExecutionResult.Success();
-
-    internal ExecutionResult VisitImportRequire(Stmt.ImportRequire importReq) => ExecuteImportRequire(importReq);
-
-    internal ExecutionResult VisitExport(Stmt.Export exportStmt) => ExecuteExport(exportStmt);
-
-    internal ExecutionResult VisitDirective(Stmt.Directive directive) =>
-        // Directives are processed at the start of interpretation for their side effects (strict mode)
-        // When encountered during execution, they are a no-op
-        ExecutionResult.Success();
-
-    internal ExecutionResult VisitDeclareModule(Stmt.DeclareModule declareModule) =>
-        // Module/global augmentations and ambient declarations are type-only
-        // No runtime effect - types were merged during type checking
-        ExecutionResult.Success();
-
-    internal ExecutionResult VisitDeclareGlobal(Stmt.DeclareGlobal declareGlobal) =>
-        // Module/global augmentations and ambient declarations are type-only
-        // No runtime effect - types were merged during type checking
-        ExecutionResult.Success();
-
-    internal ExecutionResult VisitUsing(Stmt.Using usingStmt) => ExecuteUsingDeclaration(usingStmt);
 
 }
